Move to spacy.analysis
